Propagated 0.8.2 modifs.

+++ src/tone_reproductor.h
@@ -79,7 +79,7 @@ public:
 	// Return adapted luminance from world to display
 	float adapt_luminance(float world_luminance) const
 	{
-		return powf(world_luminance*M_PI*0.0001f,alpha_wa_over_alpha_da) * term2;
+		return std::pow((float)(world_luminance*M_PI*0.0001f),alpha_wa_over_alpha_da) * term2;
 	}
 
 	// Convert from xyY color system to RGB
